requirements.txt
Django==4.2.17
mysqlclient==2.2.4
djangorestframework==3.15.1
gunicorn==21.2.0